move the InitialCondition tests to the tests_inifield.py
